The http client core becomes almide-rt-core: one source spliced into the template and linked by the embedded host
